split extraction_tools into various submodules
